pyproject.toml
[project]
name = "version0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "boto3>=1.39.4",
    "fake-useragent>=2.2.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "webdriver-manager>=4.0.2",
]
